better cors config
